refactor forms DRY
